Project - Beginner Level MySQL
In this project you will learn how the database and Table are created using MySQL query. This stage is known as Data Defined Language (DDL). After building the database and table, you can insert the data, manipulate the data, retrieve the information, inspect the data with the help of "SELECT", "ALTER", "UPDATE" and  "WHERE" clause. This stage is known as Data Manipulative Language (DML)